split FormController into LoginController (to test symfony security system) and TokenController that can be used by SPA to retrieve a valid Token
