+++ config/services.yaml
@@ -2,6 +2,7 @@
 # https://symfony.com/doc/current/best_practices/configuration.html#application-related-configuration
 parameters:
     locale: 'en'
+    csrf_token_parameter: '_csrf_token'
     csrf_token_id: 'thisIsMyTokenID_sdf468xcv*/;7{d5!LMM}'
 
 services:
